% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\title{drawR main canvas workbench compinent}
\usage{
canvas(
  canvasID,
  title = "Canvas",
  height = "85vh",
  width = "100\%",
  on_start = TRUE
)
}
\arguments{
\item{canvasID}{string, an unique HTML ID}

\item{title}{string, title of the canvas}

\item{height}{string, css value of initial height of the canvas, like "100vh"
for full height current window, "50vh" for half.}

\item{width}{string, css value of initial width of the canvas}

\item{on_start}{bool}
}
\value{
a HTML component to be added to a Shiny app or document
}
\description{
use this function on Shiny UI or R markdown to create the
image editing area.
}
\details{
\strong{If you are not working in Shiny or R markdown, you need to add the}
\strong{required full "Bootstrap3" javascript and CSS + latest "jquery" + "font-awesome" dependencies by yourself.}
\itemize{
\item \code{height}, css style \code{vh} is safer than \verb{\%} is not safe, unless the parent
has some defined height, "\%" will work. Otherwise, if the parent height is "auto"
or not defined, and you choose "100\%", canvas will still have 0 height.
}
}
